refactor: 프로젝트 구조를 응집성 있도록 변경
