+++ src/App.tsx
@@ -1,12 +1,17 @@
+import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { GlobalPortal, GlobalStyles } from 'tosslib';
-import { Routes } from './pages/Routes';
+import { Routes } from './Routes.tsx';
 
 export function App() {
+  const queryClient = new QueryClient();
+
   return (
     <>
       <GlobalStyles />
       <GlobalPortal.Provider>
-        <Routes />
+        <QueryClientProvider client={queryClient}>
+          <Routes />
+        </QueryClientProvider>
       </GlobalPortal.Provider>
     </>
   );
