+++ src/apis/fetchers/index.ts
@@ -1,0 +1,1 @@
+export { fetchSavingsProducts } from './savings';
